add outcome video controls. breakout components into separate files.
