+++ packages/app/src/components/UI/index.ts
@@ -2,6 +2,7 @@ import Button from './Button';
 import Card from './Card';
 import Fade from './Fade';
 import VideoPlayer from './Video';
+import LoadingSpinner from './LoadingSpinner';
 import { colors } from './globalStyles';
 
-export { Button, Card, Fade, VideoPlayer as Video, colors };
+export { Button, Card, Fade, VideoPlayer as Video, LoadingSpinner, colors };
