+++ packages/app/src/components/Vote/Outcome.tsx
@@ -1,14 +1,9 @@
 import React, { ReactElement, useState } from 'react';
 import { Text, View, StyleSheet, Alert } from 'react-native';
-import { colors, Card, Button } from '../UI';
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { OutcomeVideo } from './OutcomeVideo';
-import RadioForm, {
-    RadioButton,
-    RadioButtonInput,
-    RadioButtonLabel
-} from 'react-native-simple-radio-button';
 
+import { Card, Video } from '../UI';
+import { Poll } from './Poll';
+import { VoteForm } from './VoteForm';
 interface Props {}
 
 export const Outcome = (props: Props): ReactElement => {
@@ -45,7 +40,13 @@ export const Outcome = (props: Props): ReactElement => {
     return (
         <View style={styles.container}>
             <View style={styles.videoContainer}>
-                <OutcomeVideo uri="https://www.w3schools.com/html/mov_bbb.mp4" />
+                <Video
+                    source={{
+                        uri: 'https://www.w3schools.com/html/mov_bbb.mp4'
+                    }}
+                    style={{ alignSelf: 'center' }}
+                    controls={true}
+                />
             </View>
 
             <View style={styles.infoContainer}>
@@ -62,54 +63,12 @@ export const Outcome = (props: Props): ReactElement => {
                 </Card>
                 <View style={styles.options}>
                     {!voted ? (
-                        <View>
-                            <RadioForm formHorizontal={false} animation={true}>
-                                {radioOptions.map((obj, i) => (
-                                    <RadioButton
-                                        labelHorizontal={true}
-                                        key={i}
-                                        style={styles.radioButton}
-                                    >
-                                        <RadioButtonInput
-                                            obj={obj}
-                                            index={i}
-                                            isSelected={selectedIndex === i}
-                                            onPress={() => setSelectedIndex(i)}
-                                            borderWidth={1}
-                                            buttonSize={16}
-                                            buttonOuterSize={24}
-                                            buttonStyle={styles.radioInput}
-                                            buttonWrapStyle={{ width: 52 }}
-                                            buttonColor="#6084b3"
-                                        />
-                                        <RadioButtonLabel
-                                            obj={obj}
-                                            index={i}
-                                            labelHorizontal={true}
-                                            onPress={() => setSelectedIndex(i)}
-                                            labelStyle={styles.radioLabel}
-                                            labelWrapStyle={[
-                                                styles.radioLabelWrap,
-                                                {
-                                                    borderColor:
-                                                        selectedIndex === i
-                                                            ? '#6084b3'
-                                                            : '#f3f3f3'
-                                                }
-                                            ]}
-                                        />
-                                    </RadioButton>
-                                ))}
-                            </RadioForm>
-                            <Button
-                                small
-                                title="Vote"
-                                color="black"
-                                bgColor="white"
-                                style={styles.button}
-                                onPress={handleSubmit}
-                            />
-                        </View>
+                        <VoteForm
+                            radioOptions={radioOptions}
+                            selectedIndex={selectedIndex}
+                            setSelectedIndex={setSelectedIndex}
+                            handleSubmit={handleSubmit}
+                        />
                     ) : (
                         <View>
                             <Text style={styles.voted}>
@@ -130,35 +89,6 @@ export const Outcome = (props: Props): ReactElement => {
         </View>
     );
 };
-
-type PollProps = {
-    percent: number;
-    title: string;
-    usersChoice?: boolean;
-};
-
-export const Poll = ({
-    percent,
-    title,
-    usersChoice
-}: PollProps): ReactElement => (
-    <View style={styles.result}>
-        <Text style={styles.percent}>{percent}%</Text>
-
-        <View style={styles.poll}>
-            <View style={styles.pollText}>
-                <Text style={styles.optionTitle}>{title}</Text>
-                <Icon
-                    name="checkbox-marked-circle"
-                    size={18}
-                    color="black"
-                    style={{ marginLeft: 10, opacity: usersChoice ? 0.5 : 0 }}
-                />
-            </View>
-            <View style={[styles.pollResult, { width: `${percent}%` }]} />
-        </View>
-    </View>
-);
 
 const styles = StyleSheet.create({
     container: {
@@ -201,69 +131,5 @@ const styles = StyleSheet.create({
     voted: {
         textAlign: 'center',
         color: '#777'
-    },
-    optionTitle: {
-        color: 'rgba(0,0,0,.7)',
-        fontWeight: 'bold'
-    },
-    result: {
-        marginTop: 16,
-        flexDirection: 'row',
-        alignItems: 'center'
-    },
-    percent: {
-        fontWeight: 'bold',
-        marginRight: 4,
-        width: 48
-    },
-    radio: {
-        justifyContent: 'center',
-        marginTop: 20
-    },
-    radioButton: {
-        alignItems: 'center',
-        marginTop: 16
-    },
-    radioInput: {
-        marginRight: 8,
-        width: 48
-    },
-    radioLabelWrap: {
-        backgroundColor: '#f3f3f3',
-        flex: 1,
-        borderRadius: 6,
-        overflow: 'hidden',
-        paddingVertical: 9,
-        paddingHorizontal: 3,
-        borderWidth: 1
-    },
-    radioLabel: {
-        backgroundColor: '#f3f3f3',
-        fontWeight: 'bold',
-        color: 'rgba(0,0,0,.7)'
-    },
-    poll: {
-        flex: 1,
-        backgroundColor: '#f3f3f3',
-        overflow: 'hidden',
-        borderRadius: 6,
-        paddingVertical: 10
-    },
-    pollResult: {
-        backgroundColor: '#c7d9f1',
-        width: '75%',
-        position: 'absolute',
-        top: 0,
-        bottom: 0,
-        left: 0,
-        zIndex: -1
-    },
-    pollText: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        paddingHorizontal: 16
-    },
-    button: {
-        marginTop: 24
     }
 });
